refactoring: finalization
