+++ src/lib-usage.js
@@ -1,0 +1,3 @@
+import { CryptoGost, GostEngine } from './index';
+
+console.log(CryptoGost, GostEngine);
